requirements.txt
oauth2client
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
gunicorn==0.17.2
flask-mongoengine==0.7.0
mongoengine==0.8.4
httplib2
Flask-Mail
Flask-RESTful
PyJWT
